requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.1
asyncpg==0.29.0
gunicorn==21.2.0
SQLAlchemy==2.0.41